+++ packages/uring_api/src/uring_api/include/uring_api_capi.h
@@ -72,7 +72,8 @@ typedef struct UringApi_CAPI {
     int (*ring_submit_sendmsg_zc)(PyObject *ring, int fd, PyObject *data, PyObject *address, unsigned int flags,
                                   PyObject *user_data);
     int (*ring_submit_accept)(PyObject *ring, int fd, unsigned int flags, PyObject *user_data);
-    int (*ring_submit_accept_multishot)(PyObject *ring, int fd, unsigned int flags, PyObject *user_data);
+    int (*ring_submit_accept_multishot)(PyObject *ring, int fd, unsigned int flags, PyObject *user_data,
+                                        unsigned long long base_sequence);
     int (*ring_submit_connect)(PyObject *ring, int fd, PyObject *address, PyObject *user_data);
     int (*ring_submit_poll)(PyObject *ring, int fd, unsigned int mask, PyObject *user_data);
     int (*ring_submit_poll_multishot)(PyObject *ring, int fd, unsigned int mask, PyObject *user_data);